clean up, moved hdf5open and hdf5create code to their own code files
