+++ libhdf5/CMakeLists.txt
@@ -7,7 +7,7 @@
 
 # The source files for the HDF5 dispatch layer.
 SET(libnchdf5_SOURCES nc4hdf.c nc4info.c hdf5file.c hdf5attr.c
-hdf5dim.c hdf5grp.c hdf5type.c hdf5internal.c hdf5create.c)
+hdf5dim.c hdf5grp.c hdf5type.c hdf5internal.c hdf5create.c hdf5open.c)
 
 # Build the HDF4 dispatch layer as a library that will be included in
 # the netCDF library.
