+++ libhdf5/Makefile.am
@@ -13,8 +13,8 @@ libnetcdf4_la_CPPFLAGS = ${AM_CPPFLAGS}
 noinst_LTLIBRARIES = libnchdf5.la
 
 # The source files.
-libnchdf5_la_SOURCES = nc4hdf.c nc4info.c hdf5file.c hdf5attr.c	\
-hdf5dim.c hdf5grp.c hdf5type.c hdf5internal.c hdf5create.c
+libnchdf5_la_SOURCES = nc4hdf.c nc4info.c hdf5file.c hdf5attr.c		\
+hdf5dim.c hdf5grp.c hdf5type.c hdf5internal.c hdf5create.c hdf5open.c
 
 # Package this for cmake build.
 EXTRA_DIST = CMakeLists.txt
